+++ protected/views/site/admin.php
@@ -82,6 +82,12 @@
 
 
         </div>
+        </td>
+        </tr>
+        </table>
+
     </div>
+</div>
+</div>
 
 </div>